refactor(interpolation): move interpolation and its list, as well as type list, into `container` dir
