+++ src/CMakeLists.txt
@@ -5,12 +5,12 @@ configure_file(
 )
 
 set(HEADER_FILES
-        atom_type_lists.h
+        container/atom_type_lists.h
         eam_one_way.h
         eam_phi.h
         eam.h
-        interpolation_lists.hpp
-        interpolation_object.h
+        container/interpolation_lists.hpp
+        container/interpolation_object.h
         types.h
         utils.h
         data_structure/array_map.hpp
@@ -20,12 +20,12 @@ set(HEADER_FILES
         )
 
 set(SOURCE_FILES
-        atom_type_lists.cpp
+        container/atom_type_lists.cpp
         eam_one_way.cpp
         eam_phi.cpp
         eam.cpp
         types.cpp
-        interpolation_object.cpp
+        container/interpolation_object.cpp
         parser/funcfl_parser.cpp
         parser/parser.cpp
         parser/setfl_parser.cpp
