+++ src/eam_one_way.h
@@ -5,9 +5,9 @@
 #ifndef CRYSTAL_MD_EAM_RHO_H
 #define CRYSTAL_MD_EAM_RHO_H
 
+#include "container/interpolation_lists.hpp"
+#include "container/interpolation_object.h"
 #include "data_structure/array_map.hpp"
-#include "interpolation_lists.hpp"
-#include "interpolation_object.h"
 #include "types.h"
 #include <vector>
